requirements.txt
Flask==3.0.0
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
gunicorn
Flask-Dance[google]
Flask-Login
python-dotenv
Flask-Mail
Flask-Session
Flask-SQLAlchemy